+++ src/main/java/org/openas2/processor/storage/StorageModule.java
@@ -9,4 +9,6 @@ public interface StorageModule extends ProcessorModule
 	public static final String DO_STOREMDN = "store_mdn";
 
 	public static final String DO_ARCHIVE = "store_archive";
+
+	public static final String DO_STORE_ERROR = "store_error";
 }